WikiFilm/requirements.txt
numpy
requests
Flask==2.1.2
Whoosh==2.7.4
BeautifulSoup4
python-dateutil
nltk